requirements-local.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.0.0
python-multipart>=0.0.6
sqlalchemy[asyncio]>=2.0.0
aiomysql>=0.2.0
redis>=5.0.0
minio>=7.2.0
elasticsearch[async]>=8.11.0,<9.0.0
httpx>=0.26.0
python-json-logger>=2.0.0
python-dotenv>=1.0.0
pyyaml>=6.0
PyJWT>=2.8.0
cryptography>=42.0.0
orjson>=3.9.0
xxhash>=3.4.0
pypdf>=4.0.0
python-docx>=1.0.0
openpyxl>=3.1.0
chardet>=5.2.0
aiofiles>=23.0.0
loguru>=0.7.0
openai>=1.0.0
